refactor: split controller.py and ui_components.py into focused modules
Break two monoliths into navigable modules:

- controller.py: 1863 -> 128 lines, composes 11 mixins under
  controller_mixins/ (data, style, list_view, search, item_ops,
  selection, clipboard, preview, scroll, keyboard, misc)
- ui_components.py: 1742 -> 43-line shim, splits into 7 modules
  under ui/ (icons, detection, text, list_row, help, settings,
  preview)

The shim re-exports every public name (including the legacy
toggle_search_bar alias) so existing imports keep working.

Pure restructuring -- no behavior changes intended.
