+++ clipse_gui/controller_mixins/__init__.py
@@ -1,0 +1,31 @@
+"""Mixins composing the ClipboardHistoryController.
+
+Each mixin owns one domain (data, style, list view, search, item ops, etc.).
+The controller class assembles them all via multiple inheritance.
+"""
+
+from .clipboard_mixin import ClipboardMixin
+from .data_mixin import DataMixin
+from .item_ops_mixin import ItemOpsMixin
+from .keyboard_mixin import KeyboardMixin
+from .list_view_mixin import ListViewMixin
+from .misc_mixin import MiscMixin
+from .preview_mixin import PreviewMixin
+from .scroll_mixin import ScrollMixin
+from .search_mixin import SearchMixin
+from .selection_mixin import SelectionMixin
+from .style_mixin import StyleMixin
+
+__all__ = [
+    "ClipboardMixin",
+    "DataMixin",
+    "ItemOpsMixin",
+    "KeyboardMixin",
+    "ListViewMixin",
+    "MiscMixin",
+    "PreviewMixin",
+    "ScrollMixin",
+    "SearchMixin",
+    "SelectionMixin",
+    "StyleMixin",
+]
