+++ clipse_gui/ui/__init__.py
@@ -1,0 +1,12 @@
+"""Focused UI modules split from ui_components.py.
+
+Import the specific submodule you need:
+    from .ui.icons import create_pin_icon
+    from .ui.text import highlight_search_term
+    from .ui.list_row import create_list_row_widget
+    from .ui.help import show_help_window
+    from .ui.settings import show_settings_window
+    from .ui.preview import show_preview_window
+
+The legacy `clipse_gui.ui_components` module re-exports everything for back-compat.
+"""
